Allow LaMBO2 to optionally accept a config object at initialization (#76)
* update lambo2 to accept a constructed config at initialization and a logger

Previously LaMBO2 only accepted a path to a config file at initialization and ran hydra initialization to compile this config. To allow for users to run hydra initialization themselves (and potentially include other config parameters besides those related to LaMBO2, such as black box parameters), this updates the initialization to also optionally accept a compiled config. This also adds an optional logger to track metrics and adds the fft expansion factor as a config param, which used to be hard-coded to 2.

* make x0 and y0 consistent for black box

* update config with new field

* remove extra prints

* ruff formatting

---------

Co-authored-by: Frances Ding <dingf7@gene.com>